requirements.txt
# --- Core system ---
PyYAML==6.0.2
tqdm==4.65.0
psycopg[binary]==3.1.18

# --- Git interaction ---
GitPython==3.1.43

# --- Tree-sitter ---
tree_sitter==0.21.0
tree_sitter_languages==1.10.2

# --- Embeddings stack ---
numpy<2
sentence-transformers==2.2.2
transformers>=4.44.0

# --- Torch CUDA 12.1 Stack ---
torch==2.4.0+cu121
torchvision==0.19.0+cu121
torchaudio==2.4.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
